requirements.txt
streamlit==1.30.0
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
SpeechRecognition==3.11.0
streamlit_webrtc==0.53.0
streamlit-audiorec==0.1.3
pytube==12.1.3
numpy
